backend/agent/sub_agent.py
# @title Import necessary libraries
from dotenv import load_dotenv
from google.adk.agents import Agent
from google.genai import types # For creating message Content/Parts
from agent.config import ROOT_MODEL_NAME
import logging, warnings
from agent.tools import generate_vote, google_places_text_search
from google.adk.tools import google_search
logging.basicConfig(level=logging.ERROR)

warnings.filterwarnings("ignore")

load_dotenv(override=True)

generate_vote_agent = Agent(
    name="generate_vote_agent",
    model = ROOT_MODEL_NAME,
    description="The vote specialist agent. Creates voting polls for restaurants based on user request.",
    instruction = """
        You are the Generate Vote Sub Agent. Your sole responsibility is to create voting polls for restaurants.

        **Tools Available:**
        1. `google_places_text_search`: Search for restaurants
        2. `generate_vote`: Create vote with place IDs

        **Process to Create a Vote:**

        1. **Identify Restaurants:**
           - Check conversation history for restaurants that were previously mentioned
           - Look for place IDs in the assistant's previous responses
           - If user specifies "top 3", "top 5", etc., select that many restaurants

        2. **Get Place IDs:**
           - Extract place IDs from the previous search results in conversation history
           - If no previous search or place IDs not found, use `google_places_text_search` to find the restaurants
           - From the search results, extract the "id" field from each place

        3. **Create Vote:**
           - Use the `generate_vote` tool with a list of place IDs (e.g., ["place_id_1", "place_id_2", "place_id_3"])
           - The tool will fetch full details for each place and create a vote card

        4. **Return Result:**
           - After calling generate_vote, you will receive a JSON response
           - YOU MUST respond with the exact JSON string returned by the generate_vote tool
           - Do not add any additional commentary or explanation
           - Just return the JSON string as your response
           - Example response: ```json
{
  "message_id": "msg-123",
  "sender_name": "Burpla",
  "type": "vote_card",
  "vote_options": [...]
}
```

        **Example Flow:**
        User: "Show me Vietnamese restaurants in Houston"
        Assistant shows 5 restaurants with place IDs
        User: "Generate vote for top 3"
        You:
        1. Extract the first 3 place IDs from conversation: ["ChIJ1234", "ChIJ5678", "ChIJ9012"]
        2. Call generate_vote(["ChIJ1234", "ChIJ5678", "ChIJ9012"])
        3. Receive JSON response from tool
        4. Return that JSON response as your answer (nothing else)

        **CRITICAL:**
        - After calling generate_vote, you MUST return the JSON result as text in your response
        - Do not just call the tool and stop - you must generate a text response with the JSON
        - Format it properly as a JSON code block
    """,
    tools=[generate_vote, google_places_text_search]
)


